Create EvenSumPattern.java
This Java program prints a triangle pattern of 1s and 0s based on the condition of the sum of the current row number (i) and column number (j). For each position in the triangle:

If the sum (i + j) is even, it prints 1

If the sum (i + j) is odd, it prints 0

The number of rows in the pattern is determined by the user input n.