Create RightAlignedTriangle.java
This Java program prints a right-aligned triangle pattern of asterisks (*). The number of rows (n) is taken as input from the user.

For each row i, the program prints:

(n - i) spaces first to shift the stars to the right.

Followed by i asterisks (*) in a single line.

This creates a right-aligned triangle or a left-padded triangle pattern.