Create InvertedTrianglePattern.java
This Java program prints an inverted right-angled triangle made of asterisks (*). The number of rows (n) is taken as input from the user.

It starts with n stars in the first row.

Each subsequent row has one less star than the previous.

This continues until the last row, which has just one star.