Create HollowRectanglePattern.java
This Java program prints a hollow rectangle pattern using asterisks (*). The number of rows (n) and columns (m) are taken as input from the user.

Asterisks (*) are printed on the borders:

When i is 1 or n (first or last row)

Or when j is 1 or m (first or last column)

For all inner cells, it prints a space " ", creating a hollow effect.